{"types":"C","runs":44}

+++ problem15.java
@@ -2,8 +2,8 @@ import acm.program.*;
 
 public class problem15 extends ConsoleProgram{
 	public void run(){
-		int x = readInt("Enter the first number : ");
-		int y = readInt("Enter the second number : ");
+		double x = readDouble("Enter the first number : ");
+		double y = readDouble("Enter the second number : ");
 		double n = (x+y)/2;
 		println(n);
 	}
